% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conditions_functions.R
\name{alert_functions}
\alias{alert_functions}
\title{Définition de la sortie des fonctions en cas d'erreurs}
\usage{
alert_functions(binary = FALSE)
}
\arguments{
\item{binary}{logical. Define if the functions should return TRUE/FALSE or a
message}
}
\value{
a list of function
}
\description{
Définition de la sortie des fonctions en cas d'erreurs
}

