% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jury.R
\docType{package}
\name{jury}
\alias{jury}
\alias{jury-package}
\title{jury: Un paquet pour vérifier la validité de son jury de thèse}
\description{
Le paquet jury fournit une fonction principale qui vérifie la validité de la
composition du jury de thèse
}
\section{fonction jury_check}{

Cette fonction vérifie la validité de la composition du jury de thèse au
regard des règles édictées par l'université.
}

