% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jury_check.R
\name{jury_check}
\alias{jury_check}
\title{Vérifier la validité de la composition de son jury de thèse.}
\usage{
jury_check(data, binary = FALSE, gender_biais = FALSE)
}
\arguments{
\item{data}{un data.frame ou une matrice contenant les variables civilité, rang,
local, hdr, role}
}
\value{
retourne une erreur si data n'est pas dans un bon format ou si la
composition du jury de thèse n'est pas valide.
}
\description{
Vérifier la validité de la composition de son jury de thèse.
}
\details{
le contenu des variables civilite, rang, local, hdr et role est
insensible à la casse ou aux accent. Cependant, les valeurs que peuvent prendre ces
variables sont strictement contraintes. La variable civilité ne peut prendre que
la valeur Mr ou Mme; la variable rang ne peut prendre que les valeurs A ou B
désignant respectivement directeur.es de recherche/professeur.es ou chargé.es
de recherche/maitre.s de conférence; la variable local ne peut prendre que
les valeurs local ou extérieur, indiquant si les membres du jury sont
rattachés ou non à l'école doctorale dans laquelle est inscrit le doctorant;
La variable hdr ne peut prendre que les valeurs Oui, Non ou Équivalent,
indiquant si les membres du jury possèdent ou non le diplôme d'habilitation à
diriger les recherches; la variable role ne peut prendre que les valeurs
Directeur, Encadrant, Examinateur et Rapporteur.
}
\examples{
data(jury_example)
jury_check(jury_example)
}
\seealso{
jury_check_all
}

