% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jury_check_all.R
\name{jury_check_all}
\alias{jury_check_all}
\title{Vérifier la validité de la composition de son jury de thèse.}
\usage{
jury_check_all(data, n = NULL)
}
\arguments{
\item{data}{un data.frame ou une matrice contenant les variables civilité, rang,
local, hdr, role}
}
\value{
retourne une liste contenant l'ensemble des combinaisons de membres
du jury qui donne une composition de jury de thèse valide.
}
\description{
Vérifier la validité de la composition de son jury de thèse.
}
\details{
le contenu des variables civilité, rang, local, hdr et role est
insensible à la casse. Cependant, les valeurs que peuvent prendre ces
variables sont strictement contraintes. La variable civilité ne peut prendre que
la valeur Mr ou Mme; la variable rang ne peut prendre que les valeurs A ou B
désignant respectivement directeur.es de recherche/professeur.es ou chargé.es
de recherche/maitre.s de conférence; la variable local ne peut prendre que
les valeurs local ou extérieur, indiquant si les membres du jury sont
rattachés ou non à l'école doctorale dans laquelle est inscrit le doctorant;
La variable hdr ne peut prendre que les valeurs Oui, Non ou Équivalent,
indiquant si les membres du jury possèdent ou non le diplôme d'habilitation à
diriger les recherches; la variable role ne peut prendre que les valeurs
Directeur, Encadrant, Examinateur et Rapporteur.
}
\examples{
data(people_suggestion)
jury_check_all(people_suggestion)

}
\seealso{
jury_check
}
